fastapi_project/requirements.txt
aiofiles==22.1.0
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
appdirs==1.4.4
async-generator==1.10
async-timeout==4.0.2
attrs==22.1.0
aws-shell==0.2.2
awscli==1.27.27
beautifulsoup4==4.11.1
black==22.12.0
boto3==1.26.16
botocore==1.29.27
bs4==0.0.1
cached-property==1.5.2
cachetools==5.2.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cloudinary==1.30.0
colorama==0.4.4
commonmark==0.9.1
configobj==5.0.6
contourpy==1.0.6
cryptography==39.0.1
cycler==0.11.0
dictdiffer==0.9.0
discord.py==2.2.3
docutils==0.16
elastic-transport==8.4.0
elasticsearch==7.13.4
exceptiongroup==1.0.1
fake-useragent==0.1.13
fastapi==0.86.0
filelock==3.9.0
fonttools==4.38.0
frozenlist==1.3.3
gdown==4.6.0
gensim==4.2.0
google-api-core==2.11.0
google-api-python-client==2.73.0
google-auth==2.16.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.8.0
google-cloud-core==2.3.2
google-cloud-translate==3.9.0
google-cloud-vision==3.1.4
google-search-results==2.4.1
googleapis-common-protos==1.57.0
grpcio==1.50.0
grpcio-status==1.50.0
gunicorn==20.1.0
h11==0.14.0
httpcore==0.16.3
httplib2==0.21.0
httpx==0.23.3
idna==3.4
imagecleaner==0.4.4
ImageHash==4.3.1
imageio==2.22.3
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
JPype1==1.4.1
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
konlpy==0.6.0
loguru==0.6.0
lxml==4.9.1
MarkupSafe==2.1.1
matplotlib==3.6.2
multidict==6.0.4
mypy-extensions==0.4.3
networkx==2.8.8
notion==0.0.28
notion-client==2.0.0
numpy==1.23.4
oauthlib==3.2.2
opencv-python==4.6.0.66
opensearch-py==2.0.0
outcome==1.2.0
packaging==21.3
pathspec==0.10.3
Pillow==9.3.0
platformdirs==2.6.0
prompt-toolkit==1.0.18
proto-plus==1.22.1
protobuf==4.21.12
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.10.2
Pygments==2.13.0
PyMySQL==1.0.2
pyparsing==3.0.9
PySocks==1.7.1
pytesseract==0.3.10
python-dateutil==2.8.2
python-dotenv==0.21.0
python-multipart==0.0.6
python-slugify==7.0.0
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.4.1
PyYAML==5.4.1
requests==2.28.1
requests-aws4auth==1.1.2
requests-oauthlib==1.3.1
rfc3986==1.5.0
rsa==4.7.2
s3transfer==0.6.0
scikit-image==0.19.3
scikit-learn==1.1.3
scipy==1.9.3
selenium==4.6.0
six==1.16.0
smart-open==6.3.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.43
starlette==0.20.4
text-unidecode==1.3
threadpoolctl==3.1.0
tifffile==2022.10.10
tomli==2.0.1
tqdm==4.64.1
trio==0.22.0
trio-websocket==0.9.2
tweepy==4.12.1
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
uritemplate==4.1.1
urllib3==1.26.12
uvicorn==0.19.0
wcwidth==0.2.5
webdriver-manager==3.8.5
wsproto==1.2.0
yarl==1.9.2
